update Pill. Icons, close, colors, etc.
